EXTRA: Store passwords securely in the database
Plaintext passwords are a bad idea. For simplicity, I'm using Python's
own pbkdf2_hmac to store a salted hash, but in real life using a
library like Passlib is strongly recommended.

Since this requires a db migration that would be too hard to do in
plain SQL, the previous commit added the possibility of defining Python
migrations.

- Updated database schema & fixtures to reflect this change.